+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fix missing spec info when clicking next to a property
 - Fix Firefox open file that stopped working due to react upgrade
 - Fix issue with missing bottom error panel
+- Fixed headers in left panes (Layers list and Layer editor) to remain visible when scrolling
 - _...Add new stuff here..._
 
 ## 3.0.0